requirements.txt
django==5.0.7
djangorestframework==3.15.2
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
django-phonenumbers==1.0.1
django-phonenumber-field==8.0.0
django-filter==24.2
coverage==7.6.0
django-cors-headers==4.4.0
flake8==7.1.0
gunicorn==22.0.0
pillow==10.3.0
loguru==0.7.2
django-countries[pyuca]==7.6.1